refactor(main): splits off cruise & format in their own modules
